refactor(charcounter,collapse,datepicker,dialog): sépare classe et enregistrement du tag
