+++ packages/core/src/components/charcounter/index.ts
@@ -1,0 +1,11 @@
+import { ArCharcounter } from './charcounter.js';
+
+customElements.define('ar-charcounter', ArCharcounter);
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'ar-charcounter': ArCharcounter;
+    }
+}
+
+export { ArCharcounter };
