+++ packages/core/src/components/collapse/index.ts
@@ -1,0 +1,11 @@
+import { ArCollapse } from './collapse.js';
+
+customElements.define('ar-collapse', ArCollapse);
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'ar-collapse': ArCollapse;
+    }
+}
+
+export { ArCollapse };
