+++ packages/core/src/components/datepicker/datepicker.a11y.test.ts
@@ -7,7 +7,7 @@
  */
 import { expect, fixture, html, aTimeout } from '@open-wc/testing';
 import type { ArDatepicker } from './datepicker.js';
-import './datepicker.js';
+import './index.js';
 
 // ─── Helper ───────────────────────────────────────────────────────────────────
