+++ packages/core/src/components/datepicker/index.ts
@@ -1,0 +1,11 @@
+import { ArDatepicker } from './datepicker.js';
+
+customElements.define('ar-datepicker', ArDatepicker);
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'ar-datepicker': ArDatepicker;
+    }
+}
+
+export { ArDatepicker };
