+++ packages/core/src/components/dialog/index.ts
@@ -1,0 +1,11 @@
+import { ArDialog } from './dialog.js';
+
+customElements.define('ar-dialog', ArDialog);
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'ar-dialog': ArDialog;
+    }
+}
+
+export { ArDialog };
