+++ packages/core/src/index.ts
@@ -20,6 +20,7 @@ export { ArSpinner } from './components/spinner/spinner.js';
 export { ArPagination } from './components/pagination/pagination.js';
 export { ArStepper } from './components/stepper/stepper.js';
 export { ArStepperItem } from './components/stepper-item/stepper-item.js';
+import './components/dialog/index.js';
 export { ArDialog } from './components/dialog/dialog.js';
 export { ArDropdown } from './components/dropdown/dropdown.js';
 export { ArDropdownItem } from './components/dropdown-item/dropdown-item.js';
@@ -32,7 +33,10 @@ export { ArTab } from './components/tab/tab.js';
 export { ArTabPanel } from './components/tab-panel/tab-panel.js';
 export { ArTableSort } from './components/table-sort/table-sort.js';
 export type { TableSortType, TableSortOrder } from './components/table-sort/table-sort.js';
+import './components/charcounter/index.js';
 export { ArCharcounter } from './components/charcounter/charcounter.js';
 export type { CharcounterState } from './components/charcounter/charcounter.js';
+import './components/collapse/index.js';
 export { ArCollapse } from './components/collapse/collapse.js';
+import './components/datepicker/index.js';
 export { ArDatepicker } from './components/datepicker/datepicker.js';
